P3 SOLO/requirements.txt
torch # pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torchvision
h5py
numpy
matplotlib
scipy
tensorboard
tqdm
mmcv